Development

See merge request rakshithxaloori/proeliumx-api!165

+++ feed/signals.py
@@ -7,5 +7,5 @@ from feed.models import Post
 @receiver(post_delete, sender=Post)
 def post_delete_post(sender, instance, **kwargs):
     # Delete the clip if exists
-    if hasattr(instance, "clip"):
+    if hasattr(instance, "clip") and instance.clip is not None:
         instance.clip.delete()
